+++ app/main.py
@@ -19,6 +19,7 @@ app = FastAPI(
     title="Email Management API",
     version="0.1.0",
     description="A FastAPI to send emails",
+    openapi_version="3.1.0",
     contact={"name": "API Support", "email": "support@example.com"},
     license_info={
         "name": "MIT License",
@@ -30,9 +31,17 @@ app = FastAPI(
     root_path_in_servers=False,
     servers=[
         {
-            "url": f"{os.getenv('BASE_URL', '')}{os.getenv('ROOT_PATH', '/')}",
-            "description": "Base API server",
-        }
+            "url": os.getenv("DEV_URL", "https://dev.example.com"),
+            "description": "Development server",
+        },
+        {
+            "url": os.getenv("STAGING_URL", "https://staging.example.com"),
+            "description": "Staging server",
+        },
+        {
+            "url": os.getenv("PROD_URL", "https://api.example.com"),
+            "description": "Production server",
+        },
     ]
 )